abstract no results component
