+++ packages/admin-ui/client/pages/List/index.js
@@ -2,10 +2,9 @@
 import { jsx } from '@emotion/core';
 import { Fragment, Suspense, useRef, useState } from 'react';
 
-import { Button, IconButton } from '@arch-ui/button';
-import { InfoIcon, PlusIcon } from '@arch-ui/icons';
+import { IconButton } from '@arch-ui/button';
+import { PlusIcon } from '@arch-ui/icons';
 import { Container, FlexGroup } from '@arch-ui/layout';
-import { colors } from '@arch-ui/theme';
 import { Title } from '@arch-ui/typography';
 
 import ListTable from '../../components/ListTable';
@@ -20,6 +19,7 @@ import Pagination from './Pagination';
 import Search from './Search';
 import Management, { ManageToolbar } from './Management';
 import { MoreDropdown } from './MoreDropdown';
+import { NoResults } from './NoResults';
 import {
   useListFilter,
   useListItems,
@@ -89,6 +89,8 @@ export default function ListDetails(props: Props) {
   // we want to preload the Field components
   // so that we don't have a waterfall after the data loads
   preloadViews(fields.map(({ views }) => views && views.Cell).filter(x => x));
+
+  console.log('query', query.loading, query);
 
   return (
     <Fragment>
@@ -189,65 +191,3 @@ export default function ListDetails(props: Props) {
     </Fragment>
   );
 }
-
-// ==============================
-// No Results
-// ==============================
-
-const NoResultsWrapper = ({ children, ...props }) => (
-  <div
-    css={{
-      alignItems: 'center',
-      color: colors.N30,
-      display: 'flex',
-      flexDirection: 'column',
-      fontSize: 32,
-      justifyContent: 'center',
-      padding: '1em',
-      textAlign: 'center',
-    }}
-    {...props}
-  >
-    <InfoIcon css={{ height: 48, width: 48, marginBottom: '0.5em' }} />
-    {children}
-  </div>
-);
-
-const NoResults = ({ currentPage, filters, handleReset, itemCount, list, search }) => {
-  if (filters && filters.length) {
-    return (
-      <NoResultsWrapper>
-        No {list.plural.toLowerCase()} found matching the{' '}
-        {filters.length > 1 ? 'filters' : 'filter'}
-      </NoResultsWrapper>
-    );
-  }
-  if (search && search.length) {
-    return (
-      <NoResultsWrapper>
-        No {list.plural.toLowerCase()} found matching &ldquo;
-        {search}
-        &rdquo;
-      </NoResultsWrapper>
-    );
-  }
-
-  if (currentPage !== 1) {
-    return (
-      <NoResultsWrapper>
-        <p>
-          Not enough {list.plural.toLowerCase()} found to show page {currentPage}.
-        </p>
-        <Button variant="ghost" onClick={handleReset}>
-          Show first page
-        </Button>
-      </NoResultsWrapper>
-    );
-  }
-
-  if (itemCount === 0) {
-    return <NoResultsWrapper>No {list.plural.toLowerCase()} to display yet...</NoResultsWrapper>;
-  }
-
-  return null;
-};
